Move transactionHash to storage/utils
